Dashboard/api/requirements.txt
flask
flask-cors
pyyaml
python-dotenv
requests
